+++ src/worker/_encode.js
@@ -1,9 +1,10 @@
 // mongo has some opinions about what characters are allowed as keys and ids.
 //https://stackoverflow.com/questions/12397118/mongodb-dot-in-key-name/30254815#30254815
+const specialChar = /[\\\.$]/;
 
 const encodeStr = function(str) {
   if (typeof str !== 'string') {
-    str = ''
+    str = '';
   }
   return str
     .replace(/\\/g, '\\\\')
@@ -11,48 +12,45 @@ const encodeStr = function(str) {
     .replace(/\./g, '\\u002e');
 };
 
+const encodeObj = function(obj) {
+  let keys = Object.keys(obj);
+  for(let i = 0; i < keys.length; i += 1) {
+    if (specialChar.test(keys[i]) === true) {
+      let str = encodeStr(keys[i]);
+      if (str !== keys[i]) {
+        obj[str] = obj[keys[i]];
+        delete obj[keys[i]];
+      }
+    }
+  }
+  return obj;
+};
+
 //tables & infoboxes & citations could potentially have unsafe keys
 const encodeData = function(data) {
   data = data || {};
-  //encode keys in infoboxes
-  if (data.infoboxes && data.infoboxes.length > 0) {
-    data.infoboxes.forEach(info => {
-      let keys = Object.keys(info);
-      keys.forEach(k => {
-        if (k !== encodeStr(k)) {
-          info[encodeStr(k)] = info[k];
-          delete info[k];
-        }
-      });
-    });
-  }
+
   //encode keys in citations
   if (data.citations && data.citations.length > 0) {
-    data.citations.forEach(obj => {
-      let keys = Object.keys(obj);
-      keys.forEach(k => {
-        if (k !== encodeStr(k)) {
-          obj[encodeStr(k)] = obj[k];
-          delete obj[k];
-        }
-      });
-    });
+    data.citations.map(obj => encodeObj(obj));
   }
   //cleanup table-keys
   if (data.sections && data.sections.length > 0) {
     data.sections.forEach(o => {
-      if (o.tables && o.tables.length > 0) {
-        o.tables.forEach(table => {
-          table.forEach(row => {
-            let keys = Object.keys(row);
-            keys.forEach(k => {
-              if (k !== encodeStr(k)) {
-                row[encodeStr(k)] = row[k];
-                delete row[k];
-              }
-            });
-          });
+      //encode keys in templates
+      if (o.templates) {
+        o.templates = o.templates.map(tmpl => {
+          tmpl = encodeObj(tmpl);
+          //infoboxes have their stuff here
+          if (tmpl.data) {
+            tmpl.data = encodeObj(tmpl.data);
+          }
+          return tmpl;
         });
+      }
+
+      if (o.tables && o.tables.length > 0) {
+        o.tables = o.tables.map(table => encodeObj(table));
       }
     });
   }
@@ -62,4 +60,4 @@ const encodeData = function(data) {
 module.exports = {
   encodeData: encodeData,
   encodeStr: encodeStr
-}
+};